--- ui/lib/kv/addon/routes/list-root.js
@@ -1,6 +1,0 @@
-/**
- * Copyright (c) HashiCorp, Inc.
- * SPDX-License-Identifier: MPL-2.0
- */
-
-export { default } from './list';
