+++ ui/lib/kv/addon/routes/secret/details.js
@@ -26,7 +26,7 @@ export default class KvSecretDetailsRoute extends Route {
 
     controller.breadcrumbs = [
       { label: 'secrets', route: 'secrets', linkExternal: true },
-      { label: resolvedModel.backend, route: 'list-root' },
+      { label: resolvedModel.backend, route: 'list' },
       { label: resolvedModel.path },
     ];
   }
