+++ ui/lib/kv/addon/routes/secret/edit.js
@@ -21,7 +21,7 @@ export default class KvSecretEditRoute extends Route {
 
     controller.breadcrumbs = [
       { label: 'secrets', route: 'secrets', linkExternal: true },
-      { label: resolvedModel.backend, route: 'secret' },
+      { label: resolvedModel.backend, route: 'list' },
       { label: resolvedModel.path, route: 'secret.details', model: resolvedModel.path },
       { label: 'edit' },
     ];
