+++ ui/tests/integration/components/kv-page-header-test.js
@@ -79,7 +79,7 @@ module('Integration | Component | kv | kv-page-header', function (hooks) {
     await render(
       hbs`<KvPageHeader @model={{this.model}} @breadcrumbs={{this.breadcrumbs}} @isEngineView={{true}} >
     <:tabLinks>
-          <LinkTo @route="secrets" data-test-secrets-tab="Secrets">Secrets</LinkTo>
+          <LinkTo @route="list" data-test-secrets-tab="Secrets">Secrets</LinkTo>
           <LinkTo @route="configuration" data-test-secrets-tab="Configuration">Configuration</LinkTo>
   </:tabLinks>
   </KvPageHeader>
